requirements.txt
starlette
sqlalchemy
aiofiles
jinja2
bootstrap4
python-multipart
azure-storage-blob
python-dotenv
aiohttp